third_party/blink/web_tests/external/wpt/editing/run/fontname_1-1000-expected.txt

This is a testharness.js-based test.
Found 23 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "<p>[foo</p> <p>bar]</p>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"font-family:sans-serif\\">foo</span></p><p> <span style=\\"font-family:sans-serif\\"><span>bar</span></span> </p><p><span style=\\"font-family:sans-serif\\">baz</span></p>" but got "<p><span style=\\"font-family:sans-serif\\">foo</span></p><p> <span style=\\"font-family:sans-serif\\"><span>bar</span> </span></p><p><span style=\\"font-family:sans-serif\\">baz</span></p>"
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font face=\\"sans-serif\\">foo</font></p><p> <font face=\\"sans-serif\\"><span>bar</span></font> </p><p><font face=\\"sans-serif\\">baz</font></p>" but got "<p><font face=\\"sans-serif\\">foo</font></p><p> <font face=\\"sans-serif\\"><span>bar</span> </font></p><p><font face=\\"sans-serif\\">baz</font></p>"
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "foo<code>[bar]</code>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<code><span style=\\"font-family:sans-serif\\">bar</span></code>baz" but got "foo<code style=\\"font-family:sans-serif\\">bar</code>baz"
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "foo<kbd>[bar]</kbd>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<kbd><span style=\\"font-family:sans-serif\\">bar</span></kbd>baz" but got "foo<kbd style=\\"font-family:sans-serif\\">bar</kbd>baz"
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "foo<samp>[bar]</samp>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<samp><span style=\\"font-family:sans-serif\\">bar</span></samp>baz" but got "foo<samp style=\\"font-family:sans-serif\\">bar</samp>baz"
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "foo<tt>[bar]</tt>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<tt><span style=\\"font-family:sans-serif\\">bar</span></tt>baz" but got "foo<tt style=\\"font-family:sans-serif\\">bar</tt>baz"
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "[foo<code>bar</code>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-family:sans-serif\\">foo<code><span style=\\"font-family:sans-serif\\">bar</span></code>baz</span>" but got "<span style=\\"font-family:sans-serif\\">foo<code>bar</code>baz</span>"
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "[foo<code>bar</code>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "[foo<code>bar</code>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<font face=\\"sans-serif\\">foo<code><font face=\\"sans-serif\\">bar</font></code>baz</font>" but got "<font face=\\"sans-serif\\">foo<code>bar</code>baz</font>"
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "[foo<code>bar</code>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "[foo<kbd>bar</kbd>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-family:sans-serif\\">foo<kbd><span style=\\"font-family:sans-serif\\">bar</span></kbd>baz</span>" but got "<span style=\\"font-family:sans-serif\\">foo<kbd>bar</kbd>baz</span>"
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "[foo<kbd>bar</kbd>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "[foo<kbd>bar</kbd>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<font face=\\"sans-serif\\">foo<kbd><font face=\\"sans-serif\\">bar</font></kbd>baz</font>" but got "<font face=\\"sans-serif\\">foo<kbd>bar</kbd>baz</font>"
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "[foo<kbd>bar</kbd>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "[foo<listing>bar</listing>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "[foo<listing>bar</listing>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "[foo<pre>bar</pre>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "[foo<pre>bar</pre>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
Harness: the test ran to completion.
